pyproject.toml
[project]
name = "agent-sentinel"
version = "0.1.0"
description = "Sentinel: GitHub repository security analysis and vulnerability detection agent"
authors = [
    {name = "Author", email = "author.example@gmail.com"},
]
dependencies = [
    "fastmcp==2.9.2",
    "gitpython==3.0.6",
    "uvicorn",
    "python-dotenv",
    "requests==2.32.4",
    "bandit[toml]==1.8.6",
    "safety==3.4.0",
    "semgrep==1.86.0",
    "truffleHog==2.2.1",
    "detect-secrets==1.5.0",
    "pip-audit==2.9.0",
    "pyyaml==6.0.2",
    "dockerfile-parse==2.0.1",
    "pathspec==0.12.1",
    "packaging==25.0",
    "slither-analyzer==0.11.3",
    "tomlkit==0.13.3",
    "json-repair==0.47.6"
]
requires-python = ">=3.8"
license = {text = "MIT"}

[project.scripts]
agent-sentinel = "agent_sentinel.main:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_sentinel"]
